+++ test/search_test.rb
